Component Contract test concentrated in one module now.
