+++ test/components/checkbox.spec.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import {ClientRenderer, expect, simulate, sinon, waitFor} from 'test-drive-react';
 import {BasicDemo, demoCheckBoxText, DisabledDemo, IndeterminateDemo} from '../../demo/components/checkbox-demo';
 import {CheckBox, CheckBoxIconProps} from '../../src';
-import {assertRootElementContract} from '../contracts/root-element';
 
 const boxSVG: React.SFC<CheckBoxIconProps> = props => {
     return (
@@ -34,8 +33,6 @@ describe('<Checkbox/>', function() {
     afterEach(() => {
         clientRenderer.cleanup();
     });
-
-    assertRootElementContract(CheckBox);
 
     describe('Component / Demo test', function() {
         it('Basic demo', async function() {
