+++ test/node/ssr.node-spec.tsx
@@ -2,8 +2,7 @@ import * as React from 'react';
 import {renderToString} from 'react-dom/server';
 import {expect} from 'test-drive-react';
 import * as WixReactComponents from '../../src';
-
-const isReactComponent = (value: any) => value && value.prototype && value.prototype instanceof React.Component;
+import {isReactComponent} from '../utils/is-react-component';
 
 describe('SSR compatibility', function() {
     const libExportNames = Object.keys(WixReactComponents);
